requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.5.4=py37h1de35cc_0
asn1crypto=0.24.0=py37_0
async-timeout=3.0.1=py37_0
attrs=19.1.0=py37_1
ca-certificates=2019.5.15=1
certifi=2019.6.16=py37_1
cffi=1.12.3=py37hb5b8e2f_0
chardet=3.0.4=py37_1003
click=7.0=py37_0
cryptography=2.7=py37ha12b0ac_0
flask=1.1.1=py_0
idna=2.8=py37_0
itsdangerous=1.1.0=py37_0
jinja2=2.10.1=py37_0
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
markupsafe=1.1.1=py37h1de35cc_0
multidict=4.5.2=py37h1de35cc_0
ncurses=6.1=h0a44026_1
openssl=1.1.1c=h1de35cc_1
pip=19.2.2=py37_0
pycparser=2.19=py37_0
pyopenssl=19.0.0=py37_0
pysocks=1.7.0=py37_0
python=3.7.4=h359304d_1
readline=7.0=h1de35cc_5
redis-py=3.2.1=py_0
requests=2.22.0=py37_0
selenium=3.141.0=py37h1de35cc_0
setuptools=41.0.1=py37_0
six=1.12.0=py37_0
sqlite=3.29.0=ha441bb4_0
tk=8.6.8=ha441bb4_0
ujson=1.35=py37h1de35cc_0
urllib3=1.24.2=py37_0
werkzeug=0.15.5=py_0
wheel=0.33.4=py37_0
xz=5.2.4=h1de35cc_4
yarl=1.3.0=py37h1de35cc_0
zlib=1.2.11=h1de35cc_3
